+++ web/app/(service)/webhooks/stripe/_route.ts
@@ -1,0 +1,84 @@
+import { NextResponse } from "next/server";
+import { stripe } from "@/lib/clients";
+
+// TODO: V3 - need sprt of `express.raw({ type: "application/json" })` equivalent
+export async function POST(request: Request) {
+  let event: any = request.body;
+
+  if (!event) {
+    return NextResponse.error();
+  }
+
+  // Replace this endpoint secret with your endpoint's unique secret
+  // If you are testing with the CLI, find the secret by running 'stripe listen'
+  // If you are using an endpoint defined with the API or dashboard, look in your webhook settings
+  // at https://dashboard.stripe.com/webhooks
+  const endpointSecret = process.env.STRIPE_WEBHOOK_SECRET;
+  // Only verify the event if you have an endpoint secret defined.
+  // Otherwise use the basic event deserialized with JSON.parse
+  if (endpointSecret) {
+    // Get the signature sent by Stripe
+    const signature = request.headers.get("stripe-signature");
+    try {
+      event = stripe.webhooks.constructEvent(
+        request.body,
+        signature,
+        endpointSecret
+      );
+    } catch (err: any) {
+      console.log(`⚠️  Webhook signature verification failed.`, err.message);
+      return NextResponse.error();
+    }
+  }
+
+  let subscription;
+  let status;
+
+  // Handle the event
+  switch (event!.type) {
+    case "customer.subscription.trial_will_end":
+      subscription = event.data.object;
+      status = subscription.status;
+      console.log(`Subscription status is ${status}.`);
+      // Then define and call a method to handle the subscription trial ending.
+      // handleSubscriptionTrialEnding(subscription);
+      break;
+    case "customer.subscription.deleted":
+      subscription = event.data.object;
+      status = subscription.status;
+      console.log(`Subscription status is ${status}.`);
+      // Then define and call a method to handle the subscription deleted.
+      // handleSubscriptionDeleted(subscriptionDeleted);
+      break;
+    case "customer.subscription.created":
+      subscription = event.data.object;
+      status = subscription.status;
+      console.log(`Subscription status is ${status}.`);
+      // Then define and call a method to handle the subscription created.
+      // handleSubscriptionCreated(subscription);
+      break;
+    case "customer.subscription.updated":
+      subscription = event.data.object;
+      status = subscription.status;
+      console.log(`Subscription status is ${status}.`);
+      // Then define and call a method to handle the subscription update.
+      // handleSubscriptionUpdated(subscription);
+      break;
+    default:
+      // Unexpected event type
+      console.log(`Unhandled event type ${event.type}.`);
+  }
+
+  // Return a 200 response to acknowledge receipt of the event
+  NextResponse.json({}, { status: 200 });
+}
+
+// router.post(
+//   /**
+//    * /webhooks/stripe - defined in routes/index.ts
+//    */
+//   "/",
+//   express.raw({ type: "application/json" }),
+//   (request, response) => {
+//   }
+// );
